// typings/api.d.ts
export type Player = {
  id: number;
  name: string;
  position: 'C' | 'RW' | 'LW' | 'RD' | 'LD';
  league: number;
  team: string;
  season: number;
  gamesPlayed: number;
  timeOnIce: number;
  goals: number;
  assists: number;
  points: number;
  plusMinus: number;
  pim: number;
  ppGoals: number;
  ppAssists: number;
  ppPoints: number;
  ppTimeOnIce: number;
  shGoals: number;
  shAssists: number;
  shPoints: number;
  shTimeOnIce: number;
  fights: number;
  fightWins: number;
  fightLosses: number;
  faceoffs: number | undefined;
  faceoffWins: number | undefined;
  gwg: number | undefined;
  hits: number;
  giveaways: number;
  takeaways: number;
  shotsBlocked: number;
  shotsOnGoal: number;
  gameRating: number;
  offensiveGameRating: number;
  defensiveGameRating: number;
};

export type PlayerInfo = {
  height: number;
  id: number;
  league: number;
  season: number;
  weight: number;
  team: string;
  position: 'C' | 'LW' | 'RW' | 'LD' | 'RD';
};

export type PlayerWithAdvancedStats = Player & {
  advancedStats: {
    PDO: number;
    GF60: number;
    GA60: number;
    SF60: number;
    SA60: number;
    CF: number;
    CA: number;
    CFPct: number;
    CFPctRel: number;
    FF: number;
    FA: number;
    FFPct: number;
    FFPctRel: number;
  };
};

export type Goalie = {
  id: number;
  name: string;
  position: 'G';
  league: number;
  team: string;
  season: number;
  gamesPlayed: number;
  minutes: number;
  wins: number;
  losses: number;
  ot: number;
  shotsAgainst: number;
  saves: number;
  goalsAgainst: number;
  gaa: number;
  shutouts: number;
  savePct: number;
  gameRating: number;
};